% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table}
\alias{create_summary_table}
\title{Create a summary table of participant characteristics}
\usage{
create_summary_table(baseline_data)
}
\arguments{
\item{baseline_data}{dataframe}
}
\description{
Create a summary table of participant characteristics
}
